Added node.js facade dependency
